# backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.AuthRequest:
    properties:
      password:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  response.AuthResponse:
    properties:
      expiry:
        type: integer
      fullName:
        type: string
      redditUrlLogin:
        type: string
      resultInfo:
        $ref: '#/definitions/response.BaseResponse'
      token:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
  response.BaseResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
      success:
        type: boolean
    type: object
...
host: api.onesocial.kube.lucasgrey.dev
info:
  contact:
    email: me@jonathanhosea.dev
    name: John Nate
    url: https://jonathanhosea.dev
  description: Aggregator for all social media
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://onesocial.io
  title: One Social
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: login and retrieve JWT token
      parameters:
      - description: Auth Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.AuthResponse'
      summary: Login user
      tags:
      - accounts
  /auth/register:
    post:
      consumes:
      - application/json
      description: register new user and retrieve JWT token
      parameters:
      - description: Auth Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.AuthResponse'
      summary: Register new user
      tags:
      - accounts
...
securityDefinitions:
  Bearer Token:
    type: basic
swagger: "2.0"